pulsecorp/requirements.txt
asgiref==3.2.7
backports.zoneinfo==0.2.1
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
click==8.1.3
cryptography==37.0.2
defusedxml==0.7.1
Django==3.0.5
django-3-jet==1.0.8
editdistpy==0.1.3
feedparser==6.0.2
filelock==3.7.0
gensim==4.2.0
huggingface-hub==0.7.0
idna==2.9
importlib-metadata==4.11.4
jeepney==0.8.0
jira==3.2.0
joblib==1.1.0
keyring==23.5.1
language-detector==5.0.2
nltk==3.7
numpy==1.18.2
oauthlib==3.1.0
packaging==21.3
pandas==1.0.3
Pillow==7.1.1
psycopg2-binary==2.8.6
pycparser==2.21
pycrypto==2.6.1
pycryptodome==3.14.1
pyparsing==3.0.9
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
regex==2022.4.24
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
scikit-learn==1.0.2
scipy==1.7.3
SecretStorage==3.3.2
sgmllib3k==1.0.0
six==1.14.0
sklearn==0.0
smart-open==6.0.0
sqlparse==0.3.1
stop-words==2018.7.23
symspellpy==6.7.6
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.11.0
tqdm==4.64.0
transformers==4.19.2
typing_extensions==4.2.0
urllib3==1.25.8
zipp==3.8.0
